modified and integrated accounts module

+++ backend/coreBanking/src/main/java/com/bankingSystem/coreBanking/Service/AccountsService/AccountService.java
@@ -3,12 +3,15 @@ package com.bankingSystem.coreBanking.Service.AccountsService;
 import com.bankingSystem.coreBanking.DTO.AccountDTO;
 import com.bankingSystem.coreBanking.Entity.Accounts.Account;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface AccountService {
 
     List<Account> getAccounts(Long userId);
+
     AccountDTO createAccount(AccountDTO accountDTO);
-    Double getBalance(String accNo, String pin);
+
+    BigDecimal getBalance(String accNo, String pin);  // <-- changed from Double to BigDecimal
 
 }
